+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Global option "enable_blktrace". Runs blktrace and grabs relevant metrics.
 - Output option "append_blktrace". Appends blktrace metrics to output.
 - Output option "ignore_missing". If missing metric, "NONE" is used as value in output.
+- `x` `--cleanup-files` argument added. Cleans up intermediate files.
 
 ### Fixed
 - Fixed output parsing exception being handled properly.